dotfiles v5: new wallpaper, widget, instructions file and fixes.
Another update with a new wallpaper. Finally did code formatting in the INSTALL
file. As explained in Background History inside the README.md file, do not
expect a big update, like this one, until late in the year. I plan a version 6,
but it probably is not going to come any soon. But small updates may happen.